+++ System/Collections/Dictionary.lua
@@ -92,7 +92,7 @@ PLoop(function(_ENV)
 
     --- the dictionary stream worker, used to provide stream filter, map and
     -- etc operations on a dictionary without creating any temp caches
-    __Final__() __Sealed__() __NoSuperObject__()
+    __Final__() __Sealed__() __SuperObject__(false)
     class "DictionaryStreamWorker" (function (_ENV)
         extend "IDictionary"